requirements.txt
streamlit
openai
sqlalchemy
groq
langchain_groq
langgraph
python_dotenv
itsdangerous